\name{varSelectIP-package}
\Rdversion{1.1}
\alias{varSelectIP-package}
\docType{package}
\title{
Objective Bayes Model Selection
}
\description{
This package carries out objective Bayes model selection in probit and 
regression models.
}
\details{
\tabular{ll}{
Package: \tab varSelectIP\cr
Type: \tab Package\cr
Version: \tab 0.2-1\cr
Date: \tab 2013-10-29\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
There are now two user functions in this package: varSelectIP() and varSelectOBayeslinear(). The 
former is the original package functions used for linear and probit models.  The latter is a new function that replaces the former for linear regression models as well as provides posterior inference for regression parameters.
}
\author{
Andrew Womack and Gopal, V. and Novelo, L. L. and Casella, G.

Maintainer: Andrew Womack <ajwomack@ufl.edu>
}
\keyword{ package }
\seealso{ \link{varSelectIP} \link{varSelectOBayeslinear}}
