\name{varSelectIP-package}
\Rdversion{1.1}
\alias{varSelectIP-package}
\docType{package}
\title{
Objective Bayes Model Selection
}
\description{
This package carries out objective Bayes model selection in probit and 
regression models, where the candidate models are all models having a 
number of covariates smaller than the sample size.
}
\details{
\tabular{ll}{
Package: \tab varSelectIP\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-11-26\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
There is only one user function in this package: varSelectIP(). The 
arguments to that function define the model and the low-dimensional search 
parameters to be used.
}
\author{
Gopal, V. and Novelo, L. L. and Casella, G.

Maintainer: Gopal, V. <viknesh@stat.ufl.edu>
}
\keyword{ package }
\seealso{ \link{varSelectIP}}
